chore: annotated routing 방식으로 변경
